backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.9

# HTTP & Scraping
requests==2.32.3
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0

# Google AI (simplified - only what we need)
google-generativeai==0.8.5

# Remove heavy packages we don't really use
# google-cloud-vision (not needed)
# google-cloud-language (not needed)
# google-cloud-tasks (not needed)
google-cloud-vision==3.7.4
google-cloud-language==2.13.4
google-cloud-bigquery==3.25.0
sendgrid==6.11.0
apscheduler==3.10.4
pandas==2.2.3
plotly==5.24.1
